// packages/lego-webapp/redux/models/User.ts
export interface Achievement {
  id: number;
  identifier: string;
  updatedAt: Dateish;
  level: number;
  percentage: number;
}